pyproject.toml
[project]
name = "audio-saas"
version = "0.1.0"
description = "Audio transcription SaaS with FastAPI"
authors = [
    {name = "Luís Phillipe"}
]
requires-python = ">=3.11,<3.15"
dependencies = [
    "fastapi (>=0.135.1,<0.136.0)",
    "uvicorn[standard] (>=0.41.0,<0.42.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "openai-whisper (>=20250625,<20250626)",
    "transformers (>=5.3.0,<6.0.0)",
    "torch (>=2.10.0,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "ruff (>=0.15.4,<0.16.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "pytest (>=9.0.2,<10.0.0)"
]
